% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta_county_rel_10}
\alias{zcta_county_rel_10}
\title{2010 ZCTA to County Relationship File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44410 rows and 24 columns.}
\usage{
zcta_county_rel_10
}
\description{
This dataset provides simple relationships between the 2010 ZCTAs and
Census Counties. ZIP Code Tabulation Areas (ZCTAs) are, according to the
Census, "generalized areal representations of United States Postal Service (USPS) ZIP
Code service areas."
}
\keyword{datasets}
