% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta_tract_rel_10}
\alias{zcta_tract_rel_10}
\title{2010 ZCTA to Tract Relationship File}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 148897 rows and 25 columns.
}
\usage{
zcta_tract_rel_10
}
\description{
This dataset provides simple relationships between the 2010 ZCTAs and
Census Tracts. ZIP Code Tabulation Areas (ZCTAs) are, according to the
Census, "generalized areal representations of United States Postal Service (USPS) ZIP
Code service areas."
}
\keyword{datasets}
