% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipzcta}
\alias{zipzcta}
\title{ZIP Code to ZCTA Crosswalk Table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41270 rows and 5 columns.
}
\source{
\url{http://www.udsmapper.org/zcta-crosswalk.cfm}
}
\usage{
zipzcta
}
\description{
ZIP Code to ZCTA Crosswalk Table developed by John Snow, Inc. (JSI). The
data is for the 2014 UDS report data.
}
\keyword{datasets}
